# src/main/resources/swagger.yaml
openapi: 3.0.0
info:
  title: Swagger VolgSys - OpenAPI 3.0
  description: |
    Volgsys é um projeto de estudo de API, criando micro serviços. Este é um servidor de exemplo baseado na especificação OpenAPI 3.0. Você pode obter mais informações sobre o Swagger em [swagger.io](https://swagger.io). Agora você pode nos ajudar a melhorar a API, fazendo alterações na definição ou no código. Dessa forma, com o tempo, podemos melhorar a API em geral e expor algumas das novas funcionalidades no OAS3.

    Aqui também você pode encontrar todos os pacotes publicados para download dos repositórios:

    - Projeto criado em Java, Spring Boot, JPA, Hibernate. O mesmo está publicado nestas mesmas APIs:
      [Repositório do Backend](https://github.com/CleytonPrudencio/Volgsys)

    - Projeto Angular para consumo do backend onde se encontram as APIs:
      [Repositório do Frontend Angular](https://github.com/CleytonPrudencio/front-angular)
    
    Para mais informações ou sugestões, entre em contato conosco pelo email: ticleyton@gmail.com.

  termsOfService: http://swagger.io/terms/
  contact:
    email: ticleyton@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://petstore3.swagger.io/api/v3
security:
  - JWT: [ ]
...
paths:
...
  '/user':
    put:
      tags:
        - Users
      summary: Cadastrar usuario
      description: Descricao
      operationId: setUser
      parameters:
        - in: query
          name: Nome Completo
          schema:
            type: string
          required: true
        - in: query
          name: cpf/cnpj
          description: 'Digite somente numeros'
          schema:
            type: string
          required: true
        - in: query
          name: email
          description: 'xxx@xxx.xxx'
          schema:
            type: string
          required: true
        - in: query
          name: telefone
          description: '(DDD) XXXXX-XXXX'
          schema:
            type: string
          required: true
        - in: query
          name: Data de Nascimento
          description: 'Formato: dd-mm-yyyy'
          schema:
            type: string
          required: true
        - in: query
          name: cep
          description: 'Digite somente numeros.'
          schema:
            type: string
          required: true
        - in: query
          name: password
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringResponse'
        '400':
          description: Invalid Linha
        '404':
          description: linha not found
        '405':
          description: Validation exception
  '/user/listUser':
    get:
      tags:
        - Users
      summary: listar usuario
      description: Descricao
      operationId: listUser
      parameters:
        - in: query
          name: cpfcnpj
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listUser'
        '400':
          description: Invalid Linha
        '404':
          description: linha not found
        '405':
          description: Validation exception
components:
  schemas:
    StringResponse:
      type: string
    listUser:
      type: object
      properties:
        nome:
          type: string
          example: ""
        cpfcnpj:
          type: string
          example: ""
        email:
          type: string
          example: ""
        telefone:
          type: string
          example: ""
        dtnascimento:
          type: LocalDate
          example: ""
        cep:
          type: string
          example: ""
        rua:
          type: string
          example: ""
        tipopessoa:
          type: string
          example: ""
...